change: move playlist commands to its own cog
Move the code, add the possibility to disable cog, set the corresponding log messages and cog loading.
