+++ .env.example
@@ -28,6 +28,7 @@ LOG_COUNT="7"
 # Don't forget to add/set them when enabling the cog.
 ENABLE_TW="ON"
 ENABLE_YT="ON"
+ENABLE_DB="ON"
 
 # Misc
 CROISSANT_PATH="./rsc/croissant.gif"
